Launch BLE pairing service when wifi unconfigured or stale. Refactor ready for LED matrix pairing UI.

+++ ifupdown.go
@@ -20,7 +20,7 @@ func NewInterfaceManager(iface string) (*InterfaceManager) {
 }
 
 func (im *InterfaceManager) execCmd(cmd string) {
-	if im.cmd != nil {
+	if im.cmd != nil && im.cmd.Process != nil {
 		im.cmd.Process.Kill()
 	}
